requirements.txt
django==5.2.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
redis==7.0.1
django-redis==6.0.0
